pl2 editor: splited widget.gogo
